Implemented recommendations
