+++ internal/api/controller/tasks.go
@@ -3,7 +3,9 @@ package controller
 import (
 	"encoding/json"
 	"errors"
+	"math"
 	"net/http"
+	"os"
 	"strconv"
 	"strings"
 
@@ -15,6 +17,7 @@ import (
 
 type TasksController struct {
 	TasksRepo         repository.TasksRepository
+	ProfileRepo       repository.ProfileRepository
 	TagsRepo          repository.TagsRepository
 	RepliesRepo       repository.RepliesRepository
 	NotificationsRepo repository.NotificationsRepository
@@ -88,12 +91,20 @@ func (controller *TasksController) buildTaskWrapper(uid utils.UID, task *reposit
 }
 
 func (controller *TasksController) HandleGetTasksFeed(r *http.Request) utils.HandlerResponse {
+	var err error
+	var tasks []repository.Task
+
 	uid := utils.GetUserID(r.Context())
 
 	scope := r.FormValue("scope")
 	query := r.FormValue("query")
 
-	tasks, err := controller.TasksRepo.GetTasksFeed(scope, query, uid)
+	if scope == repository.RECOMMENDATIONS {
+		tasks, err = controller.GetRecommendations(uid)
+	} else {
+		tasks, err = controller.TasksRepo.GetTasksFeed(scope, query, uid)
+	}
+
 	if err != nil {
 		return utils.MakeHandlerResponse(http.StatusInternalServerError, utils.MakeErrorMessage(utils.SQL_ERROR), err)
 	}
@@ -254,4 +265,113 @@ func (controller *TasksController) HandleCloseTask(r *http.Request) utils.Handle
 	}
 
 	return utils.MakeHandlerResponse(http.StatusOK, struct{}{}, nil)
+}
+
+func normalizeVector(vector map[utils.UID]float32) {
+	magnitude := float32(0)
+	for _, val := range vector {
+		magnitude += val * val
+	}
+	magnitude = float32(math.Sqrt(float64(magnitude)))
+
+	for id, val := range vector {
+		vector[id] = val / magnitude
+	}
+}
+
+func buildUserRecommendationsVector(userTags []repository.TaskTagLink) map[utils.UID]float32 {
+	result := map[utils.UID]float32{}
+	uniqueTasks := map[utils.UID]struct{}{}
+
+	for _, row := range userTags {
+		result[row.TagID] += 1
+		uniqueTasks[row.TaskID] = struct{}{}
+	}
+
+	for tagID := range result {
+		result[tagID] /= float32(len(uniqueTasks))
+	}
+
+	normalizeVector(result)
+
+	return result
+}
+
+func buildTasksRecommendationsVector(tasksTags []repository.TaskTagLink) map[utils.UID]map[utils.UID]float32 {
+	result := map[utils.UID]map[utils.UID]float32{}
+	uniqueTasks := map[utils.UID]struct{}{}
+	uniqueTags := map[utils.UID]float32{}
+
+	for _, row := range tasksTags {
+		if _, contains := result[row.TaskID]; !contains {
+			result[row.TaskID] = map[utils.UID]float32{}
+		}
+
+		result[row.TaskID][row.TagID] = 0
+		uniqueTags[row.TagID] += 1
+		uniqueTasks[row.TaskID] = struct{}{}
+	}
+
+	for taskID, tagsMap := range result {
+		for tagID := range tagsMap {
+			result[taskID][tagID] = uniqueTags[tagID] / float32(len(uniqueTasks))
+		}
+
+		normalizeVector(result[taskID])
+	}
+
+	return result
+}
+
+func getRecommendedTasks(userVector map[utils.UID]float32, tasksVectors map[utils.UID]map[utils.UID]float32, threshold float32) []utils.UID {
+	result := []utils.UID{}
+	similarity := float32(0)
+
+	for taskID, tagsMap := range tasksVectors {
+		//TODO: use go coroutines
+		similarity = 0
+		for tagID, tagWeight := range tagsMap {
+			if val, ok := userVector[tagID]; ok {
+				similarity += tagWeight * val
+			}
+		}
+
+		if similarity >= threshold {
+			result = append(result, taskID)
+		}
+	}
+	return result
+}
+
+func (controller *TasksController) GetRecommendations(userID utils.UID) ([]repository.Task, error) {
+	threshold, err := strconv.ParseFloat(os.Getenv("SIMILARITY_THRESHOLD"), 32)
+	if err != nil {
+		return nil, err
+	}
+
+	userTags, err := controller.ProfileRepo.GetLikedTags(userID)
+	if err != nil {
+		return nil, err
+	}
+
+	tasksTags, err := controller.TasksRepo.GetTasksTags(userID)
+	if err != nil {
+		return nil, err
+	}
+
+	userVector := buildUserRecommendationsVector(userTags)
+	tasksVector := buildTasksRecommendationsVector(tasksTags)
+
+	recommendedTasks := getRecommendedTasks(userVector, tasksVector, float32(threshold))
+	result := []repository.Task{}
+	for _, taskID := range recommendedTasks {
+		task, err := controller.TasksRepo.GetTask(taskID)
+		if err != nil {
+			return nil, err
+		}
+
+		result = append(result, *task)
+	}
+
+	return result, nil
 }
