+++ internal/api/repository/tasks.go
@@ -12,6 +12,7 @@ const (
 	CUSTOMER_TASKS     = "CUSTOMER"
 	DOER_TASKS         = "DOER"
 	LIKED              = "LIKED"
+	RECOMMENDATIONS    = "RECOMMENDATIONS"
 )
 
 type Task struct {
@@ -25,6 +26,7 @@ type Task struct {
 
 type TasksRepository interface {
 	GetTasksFeed(scope string, request string, userID utils.UID) ([]Task, error)
+	GetTasksTags(userID utils.UID) ([]TaskTagLink, error)
 	GetTask(taskID utils.UID) (*Task, error)
 	IsTaskLiked(userID utils.UID, taskID utils.UID) (bool, error)
 	SetTaskLike(userID utils.UID, taskID utils.UID, value bool) error
@@ -176,4 +178,36 @@ func (repo *TasksSQLRepository) CreateTask(task Task) (utils.UID, error) {
 
 func (repo *TasksSQLRepository) CloseTask(taskID utils.UID, doerID utils.UID) error {
 	return repo.SQLClient.Exec("UPDATE tasks SET doer_id = $2 WHERE task_id = $1", taskID, doerID)
+}
+
+func (repo *TasksSQLRepository) GetTasksTags(userID utils.UID) ([]TaskTagLink, error) {
+	reader, err := repo.SQLClient.Query(
+		`SELECT task_tag.task_id, task_tag.tag_id 
+		FROM likes 
+		RIGHT JOIN task_tag 
+		ON likes.task_id = task_tag.task_id 
+		AND likes.user_id = $1 AND likes.active = true 
+		WHERE likes.user_id IS NULL`, userID,
+	)
+
+	if err != nil {
+		return nil, err
+	}
+	defer reader.Close()
+
+	result := []TaskTagLink{}
+	row := TaskTagLink{}
+	for {
+		ok, err := reader.NextRow(&row.TaskID, &row.TagID)
+		if err != nil {
+			return nil, err
+		}
+		if !ok {
+			break
+		}
+
+		result = append(result, row)
+	}
+
+	return result, nil
 }
